refactor: removeUserFromRole
